Add validation in update user

+++ server/src/controllers/user.js
@@ -23,27 +23,18 @@ export async function getAllUsers(req, res) {
 }
 
 /**
- * Get user by ID
- */
-export async function getUserById(req, res, userId) {
-  try {
-    const user = await userService.getUserById(userId);
-
-    res.writeHead(StatusCodes.OK, { "Content-Type": "application/json" });
-    res.end(JSON.stringify(user));
-  } catch (error) {
-    res.writeHead(StatusCodes.NOT_FOUND, {
-      "Content-Type": "application/json",
-    });
-    res.end(JSON.stringify({ error: error.message }));
-  }
-}
-
-/**
  * Update user by ID
  */
 export async function updateUser(req, res, userId) {
   try {
+    const existingUser=await userService.getUserById(userId);
+
+    if (!existingUser) {
+      res.writeHead(StatusCodes.NOT_FOUND, { "Content-Type": "application/json" });
+      res.end(JSON.stringify({ message: "User not found" }));
+      return;
+    }
+    
     const body = await parseRequestBody(req);
     const user = await userService.updateUser(userId, body);
 
@@ -62,6 +53,14 @@ export async function updateUser(req, res, userId) {
  */
 export async function deleteUser(req, res, userId) {
   try {
+    const existingUser=await userService.getUserById(userId);
+
+    if (!existingUser) {
+      res.writeHead(StatusCodes.NOT_FOUND, { "Content-Type": "application/json" });
+      res.end(JSON.stringify({ message: "User not found" }));
+      return;
+    }
+
     const message = await userService.deleteUser(userId);
 
     res.writeHead(StatusCodes.OK, { "Content-Type": "application/json" });
